remove UnrealToolOutput in favor of Nuke's own Tool delegate. UnrealToolOutput was a short sighted design decision.
