+++ src/Nuke.Unreal/UnrealBuild.Targets.cs
@@ -64,13 +64,12 @@ namespace Nuke.Unreal
             .Description("Generate project files for the default IDE of the current platform (Visual Studio or XCode)")
             .Executes(() =>
             {
-                Unreal.BuildTool(
-                    GetEngineVersionFromProject(),
+                Unreal.BuildTool(GetEngineVersionFromProject())(
                     "-projectfiles"
                     + $" -project=\"{ProjectPath}\""
                     + " -game -progress"
                     + UbtArgs.AppendAsArguments()
-                ).Run();
+                );
             });
 
         public virtual Target BuildEditor => _ => _
@@ -78,12 +77,11 @@ namespace Nuke.Unreal
             .Executes(() =>
             {
                 var platform = Unreal.GetDefaultPlatform();
-                Unreal.BuildTool(
-                    GetEngineVersionFromProject(),
+                Unreal.BuildTool(GetEngineVersionFromProject())(
                     $"{ProjectName}Editor {platform} Development"
                     + $" -Project=\"{ProjectPath}\""
                     + UbtArgs.AppendAsArguments()
-                ).Run();
+                );
             });
 
         public virtual Target Build => _ => _
@@ -100,12 +98,11 @@ namespace Nuke.Unreal
 
                     Log.Information($"{config} ran in {targetType}:");
                     var targetSuffix = targetType == UnreaTargetType.Game ? "" : targetType.ToString();
-                    Unreal.BuildTool(
-                        GetEngineVersionFromProject(),
+                    Unreal.BuildTool(GetEngineVersionFromProject())(
                         $"{ProjectName}{targetSuffix} {Platform} {config}"
                         + $" -Project=\"{ProjectPath}\""
                         + UbtArgs.AppendAsArguments()
-                    ).Run();
+                    );
                 });
             });
 
@@ -148,22 +145,21 @@ namespace Nuke.Unreal
                 configCombination.ForEach(combination =>
                 {
                     var (config, textureMode) = combination;
-                    Unreal.AutomationToolBatch(
-                        GetEngineVersionFromProject(),
-                        "BuildCookRun"
-                        + $" -ScriptsForProject=\"{ProjectPath}\""
-                        + $" -project=\"{ProjectPath}\""
-                        + $" -targetplatform={Platform}"
-                        + $" -platform={Platform}"
-                        + $" -clientconfig={config}"
-                        + " -ue4exe=UE4Editor-Cmd.exe"
-                        + " -cook"
-                        + (isAndroidPlatform ? $" -cookflavor={textureMode}" : "")
-                        + CookArguments.AppendAsArguments()
-                        + UatArgs.AppendAsArguments()
-                    )
-                    .WithWorkingDir(UnrealEnginePath)
-                    .Run();
+                    Unreal.AutomationTool(GetEngineVersionFromProject())(
+                        arguments:
+                            "BuildCookRun"
+                            + $" -ScriptsForProject=\"{ProjectPath}\""
+                            + $" -project=\"{ProjectPath}\""
+                            + $" -targetplatform={Platform}"
+                            + $" -platform={Platform}"
+                            + $" -clientconfig={config}"
+                            + " -ue4exe=UE4Editor-Cmd.exe"
+                            + " -cook"
+                            + (isAndroidPlatform ? $" -cookflavor={textureMode}" : "")
+                            + CookArguments.AppendAsArguments()
+                            + UatArgs.AppendAsArguments(),
+                        workingDirectory: UnrealEnginePath
+                    );
                 });
             });
         
@@ -171,6 +167,7 @@ namespace Nuke.Unreal
             .Description(
                 "Discover other C# projects which may contain additional Nuke targets, and add them to the main build project."
                 + " However after discovery they still need to be added to the main Build class."
+                // TODO: add them automatically via Roslyn
             )
             .Executes(() =>
             {
